backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
langchain==0.0.350
langchain-community==0.0.6
langchain-openai==0.0.2
openai==1.6.1
faiss-cpu==1.7.4
pypdf2==3.0.1
python-dotenv==1.0.0
tiktoken==0.5.2
aiofiles==23.2.1
